introduce sonarlint

+++ 15-repeated-string/src/main/java/RepeatedString.java
@@ -1,7 +1,7 @@
 public class RepeatedString {
     static long repeatedString(String s, long n) {
         // if I remove all the 'a', the number of 'a' in the original is the length of the orig - the new without 'a'
-        long numberOfAInOriginalString = s.length() - s.replace("a", "").length();
+        long numberOfAInOriginalString = (long) s.length() - s.replace("a", "").length();
 
         // in the sample abaabaabaa looks like this
         // aba | aba \ aba\ a
@@ -10,7 +10,7 @@ public class RepeatedString {
         // plus the part of the remaining string
         long partIndex = n % s.length();
         String lastString = s.substring(0, (int) partIndex);
-        long numberOfAInLastString = lastString.length() - lastString.replace("a", "").length();
+        long numberOfAInLastString = (long) lastString.length() - lastString.replace("a", "").length();
 
         return (n / s.length()) * numberOfAInOriginalString + numberOfAInLastString;
     }
